refactor: CameraBounds 데이터들을 DB에서 필터링된 데이터를 호출하는 방식으로 변경
