+++ app/src/main/java/com/example/gagyeboost/ui/map/MapFragment.kt
@@ -19,7 +19,6 @@ import com.google.android.gms.maps.GoogleMap
 import com.google.android.gms.maps.OnMapReadyCallback
 import com.google.android.gms.maps.model.BitmapDescriptorFactory
 import com.google.android.gms.maps.model.CameraPosition
-import com.google.android.gms.maps.model.LatLng
 import com.google.android.gms.maps.model.MarkerOptions
 import com.google.android.material.bottomsheet.BottomSheetBehavior
 import com.google.android.material.bottomsheet.BottomSheetDialog
@@ -48,7 +47,7 @@ class MapFragment : BaseFragment<FragmentMapBinding>(R.layout.fragment_map), OnM
         moveCameraToUser()
         if (::googleMap.isInitialized) {
             googleMap.projection.visibleRegion.latLngBounds.run {
-                (clusterManager.renderer as MyClusterRenderer).resizeBound(
+                viewModel.resizeBound(
                     southwest.latitude,
                     northeast.latitude,
                     southwest.longitude,
@@ -167,7 +166,7 @@ class MapFragment : BaseFragment<FragmentMapBinding>(R.layout.fragment_map), OnM
 
         googleMap.setOnCameraMoveListener {
             googleMap.projection.visibleRegion.latLngBounds.run {
-                (clusterManager.renderer as MyClusterRenderer).resizeBound(
+                viewModel.resizeBound(
                     southwest.latitude,
                     northeast.latitude,
                     southwest.longitude,
@@ -183,7 +182,7 @@ class MapFragment : BaseFragment<FragmentMapBinding>(R.layout.fragment_map), OnM
             addItem(markerMap)
         }
 
-        (clusterManager.renderer as MyClusterRenderer).markerBound.observe(viewLifecycleOwner) {
+        viewModel.markerBound.observe(viewLifecycleOwner) {
             val markerMap = viewModel.hashMapToMarkerMap(viewModel.dataMap.value ?: HashMap())
             addItem(markerMap)
         }
@@ -192,7 +191,6 @@ class MapFragment : BaseFragment<FragmentMapBinding>(R.layout.fragment_map), OnM
     private fun addItem(markerMap: HashMap<kotlin.Pair<Double, Double>, kotlin.Pair<String, String>>) {
         googleMap.clear()
         clusterManager.clearItems()
-
         markerMap.forEach { (latLng, addrMoney) ->
             val offsetItem =
                 MyItem(
@@ -201,10 +199,7 @@ class MapFragment : BaseFragment<FragmentMapBinding>(R.layout.fragment_map), OnM
                     addrMoney.first,
                     addrMoney.second
                 )
-            (clusterManager.renderer as MyClusterRenderer).addInBoundMarker(
-                clusterManager,
-                offsetItem
-            )
+            clusterManager.addItem(offsetItem)
         }
         clusterManager.cluster()
     }
